Add a menu for toggling horizontal and vertical snapping
When adjusting monitor positions in relation to each other it is
sometimes useful to be able to use very small increments of positioning.
With the current function, the snapping is always turned on and the
threshold for snapping might be too large to allow the flexibility
required for the optimal monitor positioning scheme.

With this change, it is possible to to adjust monitors in relation to
each other with far more presicion, while still keeping both sensible
defaults (all snapping turned ON).

This is the first time I'm writing code for Gtk, so I just read the
relevant documentation, and copied how menus are added. I don't think
these options would fit well under other menus in the GUI, but I'm not
sure.